feat: NavigationBarState 구현으로 중복 코드 최소화함
